requirements.txt
android-backup==0.2.0
appdirs==1.4.4
attrs==19.3.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
construct==2.10.56
croniter==0.3.34
cryptography==2.9.2
docutils==0.16
idna==2.10
ifaddr==0.1.7
importlib-metadata==1.7.0
jeepney==0.4.3
keyring==21.4.0
natsort==7.0.1
netifaces==0.10.9
packaging==20.4
pkginfo==1.5.0.1
pycparser==2.20
pycryptodome==3.9.8
Pygments==2.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-miio==0.5.3
pytuyapi-ipc==1.0.3
pytz==2019.3
readme-renderer==26.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
SecretStorage==3.1.2
six==1.15.0
tqdm==4.49.0
twine==3.2.0
urllib3==1.25.10
webencodings==0.5.1
zeroconf==0.25.1
zipp==3.1.0
